Added zip_code validation on User model
The validation uses Regex to verify the data is in the correct format.
However, it does not verify the zipcode is that of an actual location.
For example 00000 would pass the validation, but is not an actual zip
code. A more robust solution would involve adding a new gem or using an
api to verify the zip code but I'd rather get feedback before doing
something more involved.