Scale graph x-axis by absolute time, not by reading time
Before, the x axis of the temperature graph was set by the times of the readings, which
meant that if there were gaps in the readings, the nighttime markers would be out of sync
with the graph. Using momentjs makes it easier to manipulate dates in javascript, which
helps scale and label the graph according to absolute date -- the graph will always represent
the previous 168 hours and any readings that took place during that time. The nighttime
markers should now be accurate, even when there are readings missing.