Return user's current_temp as a numeric value
This fixes the sorting of the Current Temperature column of the violations table.
In order to accomplish this, we also added a separate method that returns the value
as a string with a degree symbol or N/A. We also cache the current_temp so that
we don't have to query the database every time we want it.